requirement.txt
========= pip install =======
numpy
scipy
torch==1.13.0
hydra-core==1.2.0
gym==0.21.0
torch-dct
wandb
trimesh


===== manually install =======
IsaacGym
IsaacGymEnvs

